# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Resume Parser API
  version: 1.0.0
  description: REST API for resume upload, parsing, candidate search, comparison, ATS scoring, analytics, billing, and API keys.
servers:
  - url: https://your-vercel-domain.vercel.app
security:
  - bearerAuth: []
paths:
  /api/resumes/upload:
    post:
      summary: Upload and parse a resume
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Parsed resume or duplicate result
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer